Add libquadmath
This was a massive pain to get working for two reasons.
Firstly, unlike with decimal float where compiler support for the types and intrinsics for basic operations are grouped behind a single gcc flag, type support for binary128 fp was enabled by default but the required intrinsics are shipped seperately for both gcc and llvm, and both are a pain to build.
With libgcc's soft-fp, it is built as part of gcc using a cross compiler for the target but since everything is built with cosmocc, that compiler(interally called xgcc) requires cosmocc itself to be called first afaik and currently just causes lots of errors because of missing headers which I was unable to solve completely but I assume can be done.
Then with compiler-rt, I initally tried to use the version that comes with cosmo but it is quite old and has the quad float functions disabled. Even if they were enabled they are hardcoded to use long float which gets detected as 80 bits on x86 and so disables them anyway. This means that I can't even use the official 17.0.3 release as it shares this limitation which was finally solved 5 days ago for the in development 18.0.0 release which does not even have an release candidates yet.

The second issue was the same xgcc issue but now for libquadmath as it also ships with the gcc source code. This time I managed to extract just enough of it to the its internal Makefile directly. This was probably also possible for libgcc but I cannot be bothered trying now that I have a working setup.