Start adding neovim and libuv
I tried this because of a request but turns out neovim needs libuv which is a massive mess of os ifdefs so I don't think it is worth fixing.
I have a partial diff for enabling most linux ifdefs but there are many recent syscalls that cosmo doesn't polyfill being used, some have fallbacks but some do not so perhaps are just unused if not supported.